# perf: optimize label sanitization